deleted problematic employee test files and created new one
